+++ sprint-mission-5/src/controllers/products/delete.product.comment.ts
@@ -1,7 +1,16 @@
 import prisma from "../../lib/prisma.js";
+import createError from "http-errors";
+import type { NextFunction, Request, Response } from "express";
 
-const deleteProductComment = async (req, res, next) => {
+const deleteProductComment = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   const reqId = Number(req.params.id);
+  if (!req.user) {
+    return next(createError(401, "Unauthorized"));
+  }
 
   try {
     await prisma.comment.delete({
